+++ src/cmd/ksh93/edit/emacs.c
@@ -614,7 +614,7 @@ process:
 	tty_cooked(ERRIO);
 	if(ed->e_nlist)
 		ed->e_nlist = 0;
-	stkset(sh.stk,ed->e_stkptr,(size_t)ed->e_stkoff);
+	stkset(sh.stk,ed->e_stkptr,ed->e_stkoff);
 	if(c == '\n')
 	{
 		out[eol++] = '\n';
